Boards: My boards endpoint
This allows the logged in user to `GET /boards/me` all of their boards.
I'm not sure if this is the right place for this sort of protected
endpoint. It might be more wise to put it at `GET /users/me/boards` so
the dependency of being logged in is more apparent.